+++ lib/libmf/ffi.rb
@@ -55,7 +55,9 @@ module Libmf
     attach_function :mf_load_model, [:string], Model.by_ref
     attach_function :mf_destroy_model, [Model.by_ref], :void
     attach_function :mf_train, [Problem.by_ref, Parameter.by_value], Model.by_ref
+    attach_function :mf_train_on_disk, [:string, Parameter.by_value], Model.by_ref
     attach_function :mf_train_with_validation, [Problem.by_ref, Problem.by_ref, Parameter.by_value], Model.by_ref
+    attach_function :mf_train_with_validation_on_disk, [:string, :string, Parameter.by_value], Model.by_ref
     attach_function :mf_predict, [Model.by_ref, :int, :int], :float
     attach_function :mf_cross_validation, [Problem.by_ref, :int, Parameter.by_value], :double
   end