+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 0.1.4 (unreleased)
 
+- Changed to BSD 3-Clause license to match LIBMF
 - Added support for reading data directly from files
 - Added `:numo` option to `p_factors` and `q_factors`
 - Improved performance of loading data by 5x
-- Changed to BSD 3-Clause license to match LIBMF
 
 ## 0.1.3 (2019-11-07)